requirements.txt
flask
flask-cors
PyJWT
flasgger